DB2: Fix: DataBind/Read of native LocalDates
