+++ ebean-core/src/main/java/io/ebeaninternal/server/type/ScalarTypeLocalDate.java
@@ -16,7 +16,7 @@ import java.time.ZonedDateTime;
 /**
  * ScalarType for java.time.LocalDate. This maps to a JDBC Date.
  */
-final class ScalarTypeLocalDate extends ScalarTypeBaseDate<LocalDate> {
+class ScalarTypeLocalDate extends ScalarTypeBaseDate<LocalDate> {
 
   ScalarTypeLocalDate(JsonConfig.Date mode) {
     super(mode, LocalDate.class, false, Types.DATE);
@@ -63,19 +63,4 @@ final class ScalarTypeLocalDate extends ScalarTypeBaseDate<LocalDate> {
     }
     return (LocalDate) value;
   }
-
-  @Override
-  public void bind(DataBinder binder, LocalDate value) throws SQLException {
-    if (value == null) {
-      binder.setNull(Types.DATE);
-    } else {
-      binder.setObject(value);
-    }
-  }
-
-  @Override
-  public LocalDate read(DataReader reader) throws SQLException {
-    return reader.getObject(LocalDate.class);
-  }
-
 }
